// src/timesheet-web/src/app/api/models/settings-standup-entry-create-request.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface SettingsStandupEntryCreateRequest {
  backlogCode?: string | null;
  backlogId?: number | null;
  deadline?: string | null;
  description?: string | null;
  section?: string | null;
  status?: string | null;
  taskText?: string | null;
  workDate?: string;
}